% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-appendix-hyperlinks.R
\name{add_appendix_hyperlinks}
\alias{add_appendix_hyperlinks}
\title{Change appendix refs to hyperlinks and match the references with their
actual appendix letter, adding it into the text so that it appears
normally in the document text}
\usage{
add_appendix_hyperlinks(x)
}
\arguments{
\item{x}{The LaTeX code as a vector of character strings, one for each line}
}
\value{
The modified LaTeX code as a vector of character strings, one for
each line
}
\description{
Change appendix refs to hyperlinks and match the references with their
actual appendix letter, adding it into the text so that it appears
normally in the document text
}
