% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-references-phantom.R
\name{add_references_phantom}
\alias{add_references_phantom}
\title{Add phantomsection latex macro after the references declaration so that
the TOC link for it works correctly}
\usage{
add_references_phantom(x)
}
\arguments{
\item{x}{The LaTeX code as a vector of character strings, one for each line}
}
\value{
The modified LaTeX code as a vector of character strings, one for
each line
}
\description{
Add phantomsection latex macro after the references declaration so that
the TOC link for it works correctly
}
\details{
This is another item that failed when we changed to lualatex. Also see
\code{\link[=add_appendix_subsection_refs]{add_appendix_subsection_refs()}} and \code{\link[=fix_table_refs]{fix_table_refs()}} which were required
to fix issues with lualatex.
}
