% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{This function will clean the current working directory of remnants of
previous builds.}
\usage{
clean(
  ...,
  book = FALSE,
  book_dir = "_book",
  knitr_figures_dir = "knitr-figs-pdf",
  knitr_cache_dir = "knitr-cache-pdf"
)
}
\arguments{
\item{...}{Optional quoted or unquoted names of chunks that should be
deleted.The chunk name(s) will be used to delete only the objects
relating to them from the \code{knitr_figures_dir} and \code{knitr-cache}.as well
as all the LaTeX and intermediate RMD and MD files. For example if there
is a file called 'example-1.png' in the \code{knitr_figures_dir} you would run
\code{clean(example)} which will remove all outputs of the 'example' knitr chunk.
If this argument is not used, All files will be deleted from
\code{knitr_figures_dir} and \code{knitr_cache_dir}
\code{knitr_figures_dir} and \code{knitr_cache_dir}}

\item{book}{Logical. If \code{TRUE}, delete the '_book' directory recursively}

\item{book_dir}{The name of the directory in which the output '\emph{.pdf' or
'}.docx' documents reside}

\item{knitr_figures_dir}{Directory where the knitr-generated
figures reside}

\item{knitr_cache_dir}{Directory where the knitr cached chunk
databases reside}
}
\value{
Nothing
}
\description{
This function will clean the current working directory of remnants of
previous builds.
}
\details{
Clean (delete) all build output including knitr cache and knitr figures
directories from the current directory.
}
