% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape-latex-symbols.R
\name{escape_latex_symbols}
\alias{escape_latex_symbols}
\title{Search in a string for LaTeX symbols and ensure they have two backslashes before them}
\usage{
escape_latex_symbols(str)
}
\arguments{
\item{s}{A string}
}
\value{
The original string \code{str} but with two backslashes preceding all
LaTeX symbols found in the symbols list in the function (\code{gr})
}
\description{
Search in a string for LaTeX symbols and ensure they have two backslashes before them
}
