% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-table-refs.R
\name{fix_table_refs}
\alias{fix_table_refs}
\title{Add extra code lines after the caption line for each table so that
the linking references work properly in the document}
\usage{
fix_table_refs(x)
}
\arguments{
\item{x}{The LaTeX code as a vector of character strings, one for each line}
}
\value{
The modified LaTeX code as a vector of character strings, one for
each line
}
\description{
Add extra code lines after the caption line for each table so that
the linking references work properly in the document
}
\details{
When using lualatex, standard table referencing fails. This hack must be
done to make the table references correct and the links, when clicked,
take you to the correct table. See comments in the code for more details.

This is another item that failed when we changed to lualatex. Also see
\code{\link[=add_references_phantom]{add_references_phantom()}} and \code{\link[=add_appendix_subsection_refs]{add_appendix_subsection_refs()}} which were
required to fix issues with lualatex.
}
