% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotest.R
\name{gotest}
\alias{gotest}
\title{Create and change to a temporary directory and copy a skeleton
version of the document files for testing}
\usage{
gotest(
  doc_dir = NULL,
  config_fn = "_bookdown.yml",
  figures_dir = NULL,
  knitr_figures_dir = NULL
)
}
\arguments{
\item{doc_dir}{If you keep your document project in a subdirectory
of the project root directory, then supply the name of that directory here}

\item{config_fn}{The name of the bookdown YAML file. The default is
\verb{_bookdown.yml}}

\item{figures_dir}{The name of the directory in which premade figures
(typically png files) are kept}

\item{knitr_figures_dir}{The name of the directory in which knitr saves
its copies of the figures it creates from knitr chunks}
}
\value{
Nothing, but the global variable \code{goback_dr} is set for the
function \code{\link[=goback]{goback()}} to use to return to the original directory
}
\description{
Create and change to a temporary directory and copy a skeleton
version of the document files for testing
}
\details{
Extracts the source and destination filenames for copying into the
temporary directory with the function \code{\link[=gotest_doc_get_src_dest_filenames]{gotest_doc_get_src_dest_filenames()}}.
See the \emph{Debugging a figure or table, or any other Rmarkdown code}
section of the README.md file for detailed explanation.
Use the function \code{\link[=goback]{goback()}} to return from the testing directory to the
directory you were in before calling \code{\link[=gotest]{gotest()}}
}
