% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotest-doc-get-src-dest-filenames.R
\name{gotest_doc_get_src_dest_filenames}
\alias{gotest_doc_get_src_dest_filenames}
\title{Create the lists of source and destination filenames needed for copying
for the document testing}
\usage{
gotest_doc_get_src_dest_filenames(
  doc_dir = NULL,
  bookdown_lst = NULL,
  figures_dir = NULL
)
}
\arguments{
\item{doc_dir}{If you keep your document project in a subdirectory
of the project root directory, then supply the name of that directory here.
This must be the full path of the directory.}

\item{bookdown_lst}{A list of values extracted from the bookdown file as
output by \code{\link[=read_bookdown_file]{read_bookdown_file()}}}

\item{figures_dir}{The name of the directory in which premade figures
(typically png files) are kept}
}
\value{
A list of two vectors, \code{src_fns} which are the source filenames
as absolute filenames (full path) and \code{dest_fns} which are the destination
filenames as relative filenames
}
\description{
Create the lists of source and destination filenames needed for copying
for the document testing
}
\details{
Meant to be called by the wrapper function \code{\link[=gotest]{gotest()}}
}
