% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-citations.R
\name{set_citations}
\alias{set_citations}
\title{Set the 'other' language citation up so that page 2 has the correct other
language based on the value of \code{fr()}}
\usage{
set_citations(
  fn = get_index_filename(system.file("rmarkdown", "templates", "resdoc", "skeleton",
    "_bookdown.yml", package = "csasdown")),
  render_type = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fn}{The bookdown index filename, typically index.Rmd. This file
must have a YAML option called 'french' set to either 'true' or 'false'}

\item{render_type}{The render type (e.g., "resdoc_pdf", "resdoc-b_pdf").
If NULL, citation checking is skipped for non-bilingual documents.}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\description{
Set the 'other' language citation up so that page 2 has the correct other
language based on the value of \code{fr()}
}
